+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file is used to list changes made in each version of grafana.
 ## Unreleased
 
 - Add more properties to generic oath resource
+- Add warning when extra_options conflict with resource properties
 
 ## 10.2.13 - *2023-04-17*